--- src/main/java/kr/hhplus/be/server/api/cart/dto/CartResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.api.cart.dto;
-
-public record CartResponse(Long productId, String name, int amount) {}
