--- src/main/java/kr/hhplus/be/server/api/option/dto/OptionDto.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.api.option.dto;
-
-public record OptionDto(
-        String name,
-        Integer additionalCost
-) { }
