--- src/main/java/kr/hhplus/be/server/api/order/dto/OrderResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.api.order.dto;
-
-public record OrderResponse(Long id, String orderNo, int amount, String status) {}
