--- src/main/java/kr/hhplus/be/server/api/point/dto/PointResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.api.point.dto;
-
-public record PointResponse(Long userId, int point) {}
