--- src/main/java/kr/hhplus/be/server/api/product/dto/ProductDto.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.api.product.dto;
-
-
-public record ProductDto(
-        Long id,
-        String itemName,
-        String optionName,
-        double finalPrice
-) {
-}
