--- src/main/java/kr/hhplus/be/server/api/user/dto/UserResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.api.user.dto;
-
-public record UserResponse(Long id, String email, String role) {}
