+++ src/main/java/kr/hhplus/be/server/application/product/ProductService.java
@@ -1,0 +1,25 @@
+package kr.hhplus.be.server.application.product;
+
+import kr.hhplus.be.server.domain.product.Product;
+import kr.hhplus.be.server.domain.product.ProductRepository;
+import kr.hhplus.be.server.interfaces.api.product.dto.ProductDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+public class ProductService {
+    private final ProductRepository productRepository;
+
+    @Transactional(readOnly = true)
+    public List<ProductDto> getProductList() {
+        List<Product> products = productRepository.findAll();
+        return products.stream()
+                .map(Product::toDto)
+                .collect(Collectors.toList());
+    }
+}
