+++ src/main/java/kr/hhplus/be/server/domain/category/CategoryRepository.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.domain.category;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
