+++ src/main/java/kr/hhplus/be/server/domain/item/ItemRepository.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.domain.item;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
