+++ src/main/java/kr/hhplus/be/server/domain/option/OptionRepository.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.domain.option;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OptionRepository extends JpaRepository<Option, Long> {
+}
