+++ src/main/java/kr/hhplus/be/server/domain/product/Product.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.domain.product;
 
 import jakarta.persistence.*;
-import kr.hhplus.be.server.api.product.dto.ProductDto;
+import kr.hhplus.be.server.interfaces.api.product.dto.ProductDto;
 import kr.hhplus.be.server.domain.common.BaseEntity;
 import kr.hhplus.be.server.domain.item.Item;
 import kr.hhplus.be.server.domain.option.Option;
