+++ src/main/java/kr/hhplus/be/server/domain/product/ProductRepository.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.domain.product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
