+++ src/main/java/kr/hhplus/be/server/interfaces/api/inventory/dto/InventoryResponse.java
@@ -1,0 +1,3 @@
+package kr.hhplus.be.server.interfaces.api.inventory.dto;
+
+public record InventoryResponse(Long productId, int stock) {}
