+++ src/main/java/kr/hhplus/be/server/interfaces/api/item/dto/ItemDto.java
@@ -1,0 +1,14 @@
+package kr.hhplus.be.server.interfaces.api.item.dto;
+
+import kr.hhplus.be.server.domain.item.SaleStatus;
+
+import java.time.LocalDateTime;
+
+public record ItemDto (
+        String name,
+        String description,
+        SaleStatus saleStatus,
+        Integer basePrice,
+        LocalDateTime saleStartDate
+){
+}
