+++ src/main/java/kr/hhplus/be/server/interfaces/api/option/dto/OptionDto.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.interfaces.api.option.dto;
+
+public record OptionDto(
+        String name,
+        Integer additionalCost
+) { }
