+++ src/main/java/kr/hhplus/be/server/interfaces/api/order/dto/OrderResponse.java
@@ -1,0 +1,3 @@
+package kr.hhplus.be.server.interfaces.api.order.dto;
+
+public record OrderResponse(Long id, String orderNo, int amount, String status) {}
