+++ src/main/java/kr/hhplus/be/server/interfaces/api/payment/dto/PaymentResponse.java
@@ -1,0 +1,3 @@
+package kr.hhplus.be.server.interfaces.api.payment.dto;
+
+public record PaymentResponse(Long orderId, String status, int paidAmount) {}
