+++ src/main/java/kr/hhplus/be/server/interfaces/api/product/dto/ProductDto.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.interfaces.api.product.dto;
+
+
+public record ProductDto(
+        Long id,
+        String itemName,
+        String optionName,
+        double finalPrice
+) {
+}
