+++ src/main/java/kr/hhplus/be/server/interfaces/api/user/dto/UserResponse.java
@@ -1,0 +1,3 @@
+package kr.hhplus.be.server.interfaces.api.user.dto;
+
+public record UserResponse(Long id, String email, String role) {}
